refactor out WMF export, make implementation less specific
